% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Get credentials}
\usage{
get_credentials()
}
\value{
The currently set email, or \code{NULL} if no email is currently set.
}
\description{
Gets the currently set email and password registered with the BYU corpora that should be used during the current R session for queries on the corpora.
}
\examples{
get_credentials()
}

