% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_bnc.R
\name{search_bnc}
\alias{search_bnc}
\title{Search the British National Corpus (BNC)}
\usage{
search_bnc(search_terms, section = "spok", max_type = 10,
  max_per_term = 100, max_total_result = 1000)
}
\arguments{
\item{search_terms}{The search term or terms, as a vector of strings.}

\item{section}{The section of the BNC to search in, from among \code{"all"} for all sections, \code{"spok"} for the spoken section (the default), \code{"fict"} for fiction, \code{"mag"} for magazine, \code{"news"} for newspaper, \code{"non-acad"} for non-academic, \code{"acad"} for academic, and \code{"misc"} for miscellaneous.}

\item{max_type}{An integer specifying the maximum number of unique word types to return for each search string (results shown in the upper right portion of the web interface). For example, searching for nouns with the search string "[n*]" could potentially return tens of thousands of unique types, but the user may only be interested in the 100 most frequent ones.}

\item{max_per_term}{An integer specifying the maximum number of keyword-in-context (KWIC) results to return for each search string.}

\item{max_total_result}{An integer specifying the maximum number of total results to return. If only one search term is given in \code{search_terms}, this argument should be equal to or greater than the integer specified in \code{max_per_term}.}
}
\value{
A data frame.
}
\description{
Retrieve keyword-in-context results from the BNC.
}
\examples{
search_bnc("cheers")
search_bnc("cheers", section = "spok")
search_bnc(c("cheers", "bye"), section = "spok", max_per_term = 500)

}

