% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cde.R
\name{search_cde}
\alias{search_cde}
\title{Search the Corpus del Español (CDE)}
\usage{
search_cde(search_terms, section = "oral", max_type = 10,
  max_per_term = 100, max_total_result = 1000)
}
\arguments{
\item{search_terms}{The search term or terms, as a vector of strings. Documentation about part-of-speech and wildcard tags is available on the CDE.}

\item{section}{The section or sections of the CDE to search in, specified by a one-hundred-year period or periods, for example, \code{section = "1900s"} or \code{section = c("1900s", "1800s")}. Specific genres of the 1900s can be specified with \code{"oral"} for the oral section, \code{"fict"} for fiction, \code{"news"} for news, and \code{"acad"} for academic. Additionally, a combination of genres and/or one hundred year periods can be specified, for example, \code{section = c("1200s", "1500s", "1800s")} or \code{section = c("news", "acad")}. To search in all centuries, use \code{section = "all"}.}

\item{max_type}{An integer specifying the maximum number of unique word types to return for each search string (results shown in the upper right portion of the CDE). For example, searching for nouns with the search string "[n*]" could potentially return tens of thousands of unique types, but the user may only be interested in the 100 most frequent ones.}

\item{max_per_term}{An integer specifying the maximum number of keyword-in-context (KWIC) results to return for each search string.}

\item{max_total_result}{An integer specifying the maximum number of total results to return. If only one search term is given in \code{search_terms}, this argument should be equal to or greater than the integer specified in \code{max_per_term}.}
}
\value{
A data frame.
}
\description{
Search the Corpus del Español.
}
\examples{
search_coca("mujer")
search_coca(c("[mujer]", "[hombre]"), section = c("oral", "acad"))

}

