% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_coca.R
\name{search_coca}
\alias{search_coca}
\title{Search the Corpus of Contemporary American English (COCA)}
\usage{
search_coca(search_terms, section = "all", max_type = 10,
  max_per_term = 100, max_total_result = 1000)
}
\arguments{
\item{search_terms}{The search term or terms, as a vector of strings.}

\item{section}{The section of the COCA to search in, from among \code{"all"} for all sections (the default), \code{"spok"} for the spoken section, \code{"fic"} for fiction, \code{"mag"} for magazine, \code{"news"} for newspaper, \code{"acad"} for academic.}
}
\value{
A data frame.
}
\description{
Search the COCA.
}
\examples{
search_coca("hello")
search_coca("dude", section = "spok")

}

