% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_coha.R
\name{search_coha}
\alias{search_coha}
\title{Search the Corpus of Historical American English (COHA)}
\usage{
search_coha(search_terms, section = "fict", max_type = 10,
  max_per_term = 100, max_total_result = 1000)
}
\arguments{
\item{search_terms}{The search term or terms, as a vector of strings.}

\item{section}{The section or sections of the COHA to search in, from among \code{"all"} for all sections, \code{"fict"} for the fiction section (the default), \code{"mag"} for magazine, \code{"news"} for newspaper, \code{"nf"} for NF Books. Also, a specific decade can be specified by the first year in the decade, for example, \code{1920} or \code{1880}. Any combination of genres and/or decades can be specified in a vector, for example, \code{section = c("1850", "1950")} or \code{section = c("mag", "news")}.}

\item{max_type}{An integer specifying the maximum number of unique word types to return for each search string (results shown in the upper right portion of the COHA). For example, searching for nouns with the search string "[n*]" could potentially return tens of thousands of unique types, but the user may only be interested in the 100 most frequent ones.}

\item{max_per_term}{An integer specifying the maximum number of keyword-in-context (KWIC) results to return for each search string.}

\item{max_total_result}{An integer specifying the maximum number of total results to return. If only one search term is given in \code{search_terms}, this argument should be equal to or greater than the integer specified in \code{max_per_term}.}
}
\value{
A data frame.
}
\description{
Search the COHA.
}
\examples{
search_coha("erstwhile")
search_coha("erstwhile", section = "mag")
search_coha(c("erstwhile", "ere"), section = c("mag", "news"), max_per_term = 500)

}

