% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Set credentials}
\usage{
set_credentials(email, password)
}
\arguments{
\item{email}{The email as a string.}

\item{password}{The password as a string.}
}
\value{
Sets email and password in the global options of the current R session with \code{options()} and returns a message to the user.
}
\description{
Sets the email and password registered with the BYU corpora that should be be used during the current R session for queries on the corpora.
}
\details{
If users are not sure which email and password are registered, they should visit \href{http://corpus.byu.edu/login.asp}{corpus.byu.edu} in a web browser to figure it out.
}
\examples{
set_credentials("my_username@institution.edu", "my_password")
}

